Создала 2 дополнительных компонента для скрытия и раскрытия комментария
